+++ packages/appcd-default-plugins/index.js
@@ -1,7 +1,7 @@
 const fs = require('fs');
 const path = require('path');
 const { dependencies } = JSON.parse(fs.readFileSync(path.join(__dirname, 'package.json')));
-const deps = new Set(Object.keys(dependencies).filter(n => n.startsWith('appcd-plugin-')));
+const deps = new Set(Object.keys(dependencies).filter(n => n.startsWith('appcd-plugin-') || n.startsWith('@appcd/plugin-')));
 let modulesDir = __dirname;
 let lastDir = null;