+++ packages/appcd-plugin/CHANGELOG.md
@@ -6,6 +6,7 @@
    [(DAEMON-262)](https://jira.appcelerator.org/browse/DAEMON-262)
  * Added list of plugin's services to the default plugin info route.
    [(DAEMON-265)](https://jira.appcelerator.org/browse/DAEMON-265)
+ * Added `appcd-plugin` to the list of injected appcd packages into plugins.
  * Updated dependencies.
 
 # v1.1.2 (May 24, 2018)