+++ packages/appcd-plugin/CHANGELOG.md
@@ -7,6 +7,7 @@
  * Added list of plugin's services to the default plugin info route.
    [(DAEMON-265)](https://jira.appcelerator.org/browse/DAEMON-265)
  * Added `appcd-plugin` to the list of injected appcd packages into plugins.
+ * Improved debug log namespace names.
  * Updated dependencies.
 
 # v1.1.2 (May 24, 2018)